Increase bathing suit animation support
Before, the bathing suit would not display in animations that were not just walking - it would just show the player's pants and no shirt.
Now, it displays the bathing suit in all of the frames I have encountered (with some exceptions in frames that could not possibly be mapped to one of the nine bathing suit frames I have to work with).
Adds a config option to enable varying levels of animation patches.
Moves all animation-related patches to the new AnimationManager.
I will add documentation to AnimationManager in a future commit.