Fix gamepad snapping
Previously, the gamepad would not snap to the resource button - it would skip over it and go to the trash can.
I removed a couple checks for Game1.activeClickableMenu is not GameMenu, because these were always true and thus prevented the snapping from being set up. This doesn't seem to have broken anything, but it is something to remember if we have future snapping issues.