Patch Farmer.getItemCount
Changes the patch on Farmer.getItemCountInList to Farmer.getItemCount. It used to be this way, but in 7f531b89bd5cdcfb06386e380c84ca784d638511 I changed it to getItemCountInList because I misunderstood the distinction between the methods. This caused the player's resources to be double-counted in the crafting menu.