Update Night Market logic
Previously, I disabled the check that prevented this mod from applying when the player was in the Night Market because if they swam into the market through the river, they would be unable to leave the water. I re-add that check, but only apply it if they aren't swimming, and also make sure the bridge is fixed to prevent softlocks on the right island.