Fix farmhouse warps being removed
This was in the mod originally, but I removed it when I updated it to 1.6 because I didn't see why it was necessary, but apparently the warps are being removed from the farmhouse for some people. I'm guessing it's a mod conflict, but idk why a mod would do that.