Update Custom Spouse Patio Redux
Fixes an issue where the patios could disappear if the asset for the map they were on was invalidated/reloaded.
Add support for multiplayer.
Ensure the patios are cached and placed each DayStarted.
Fix a bug where the config options would not be added to GMCM if Config.EnableMod was false.
Fix a bug where if a patio's map had layers that were not in the map it was being placed in, the tiles in those layers would not be removed when the patio was removed/moved. Note: that may still happen if another override adds those layers after we cache the base tiles but before we add the patio override, but I'm not going to fix that because it is fairly unlikely.
Fix a bug where if a DecoratableLocation's flooring/wallpaper was changed after a patio was placed, the old flooring/wallpaper would be re-added after the patio was removed/moved.
Fix a bug where if two patios were placed overlapping, removing the first placed and then the second would cause the overlapped section of the first patio to remain (i.e. ensure no patio tiles are ever cached in baseSpouseAreaTiles).
Fix a bug where when reloading the save data with the patio wizard, the old patios would not be removed.