Update Longer Seasons
Updates the way the billboard texture is generated to add support for mods that modify the billboard texture. Adds many config options to facilitate this.
Switches from syncing the whole config to only syncing the necessary values so that backwards compatibility does not necessarily break when we add new config options in the future.
Adds console commands to allow setting the month and the day.
Adds CP tokens for the days per month, months in current season, and current season month.
Adds an i18n and moves GMCM config option names to the i18n to allow translation.
Adds an API.
Changes the way that WorldDate.TotalDays is calculated to assume that it is acting as an offset of today's date (this is how it is almost always used in the base game)
Adds a patch to SpecialOrder.GetDaysLeft to ensure it returns an accurate value in cases where the due date appears to be in the past due to the existence of multiple months in the season, which the game does not account for.