Update Custom Gift Limits
Adds a new 'Compatibility Mode' config option that changes the gift values stored in Friendship instances so that other mods that check those but don't check our limits will still get the correct result as to if a gift can be accepted (i.e. if we set the limit to 4 gifts per day and the player has given 3, with compatibility mode it would store 0, rather than 3, in the Friendship so that it is below the default limit of 1.
Adds an API so that other mods are able to retrieve the gift limits.
Adds a new config option for limits for characters with maxed hearts.
Changes Harmony patches to be explicit, rather than using patchall. Patchall was not working when building in release mode; I suspect that the optimizer saw that the patches were unreferenced and left them out of the dll but idk.