Update oxygen bar texture
Changes the bar to be like the base game's hp and stamina bars. The new texture was created by MithranArkanere.
Adds new config options to move the oxygen bar.
Fixes an issue where one if the ocean forage items was (O)327 instead of (O)372.
Modifies a Mountain map warp at the bottom of the lake because it warped the player out of bounds.
Adds some framework for drawing the swim shadow under the player when they're swimming outside of the bathhouse, but does not implement it.
Fixes an out of bounds error with the jumping logic.