Update Better Game Menu Integration
I was trying to get this mod to work using only BGM's IPageOverlay interface and no Harmony patches, but I'm going to just leave it be at this point in time. I'm going to immediately revert this commit, but I wanted to record that this is something I tried and to keep this a starting point if I want to come back to this in the future.