Fix issue with ocean treasure
The code for generating the treasure chest items was ripped from FishingRod.openTreasureMenuEndFunction, but it hadn't been updated since before 1.6 and something in the meantime had changed with the Ring constructor or ids, so this commit updates the code to match the latest version of openTreasureMenuEndFunction.
It also adds some error checking to hopefully prevent this issue from being visible to players if it crops up again in the future.
Also renames SwimMaps.AddWaterTiles to ReloadWaterTiles to better reflect its effect