Update jump logic
Fixes an issue where some buildings tiles over water (most notably some lily pads in the Forest river) would block the player from jumping in. Will now ignore buildings tiles if the tile is a water tile and all four adjacent tiles are water tiles too.
Now checks for buildings tiles on layers like Buildings1 and Buildings-1 before jumping into water because it is somewhat common for map makers to make bridges over water using tiles on those layers, rather than passable buildings tiles. If there is an impassable buildings tile between the player and those tiles, they will still be ignored (the thought process being if there is not an impassible tile in between, it is probably a bridge, and if there is an impassible tile, it is probably in a river or similar)
Checks for Objects at the jump location before jumping to prevent the player from clipping into bushes and stones and whatnot
Disables the SwimIndoors option because many of the locations we add are indoors, so disabling it breaks many of our own locations.